handle trying to save bad responses
